fix(app,core): compila o app shell (src-tauri) e os testes do núcleo
Oitava/nona passada do `cargo check --workspace --all-targets`: os crates
`anigo-core`, `anigo-renderer` e `anigo-mcp` compilam; os erros agora são do
`anigo-app` (src-tauri) e de código de teste do núcleo.

src-tauri
- `main.rs`: o `Builder::default()....build(...)` era uma expressão solta e o
  tratamento de erro do P1-01 fazia `match app { .. }` com uma variável `app`
  inexistente. Agora o resultado é `built` e o `match` usa `built`.
- `bridge.rs`: `crate::bridge_metrics` não existe como módulo de crate (o
  `bridge_metrics` é inline dentro de `bridge.rs`) — as duas referências passam a
  ser relativas. `generate_token` fazia `read_exact` num `Vec<u8>` (o
  `std::fs::read` devolve bytes, não arquivo): agora abre `/dev/urandom` como
  `File`. `write_response` virou genérico sobre `AsyncWrite`, porque o cliente usa
  `OwnedWriteHalf` e o servidor `WriteHalf<TcpStream>` (era o E0308 em 4 pontos).
- avisos: `save_path` só é usado no bloco Windows (`cfg_attr`),
  `open_directory_in_explorer` não usa `path` fora do Windows (idem).

anigo-core (testes, compilados só agora por causa de `--all-targets`)
- `skinning.rs`: `right_sum / right_count as f32 < 0.0` era lido como argumento
  genérico (`<` depois de `f32`); parênteses explícitos nos dois asserts.
- `skinning.rs`: `joints[0]` é `u16` e `joints::ROOT` é `u32`; a comparação
  converte explicitamente.
- `win32_stub::attach_interactive_desktop` ganhou `#[allow(dead_code)]` com a
  justificativa (a API é igual nos dois `cfg`).

+++ crates/anigo-core/src/skinning.rs
@@ -743,12 +743,12 @@ mod tests {
         }
         assert!(left_count > 100 && right_count > 100, "{left_count}/{right_count}");
         assert!(
-            left_sum / left_count as f32 > 0.0,
+            left_sum / (left_count as f32) > 0.0,
             "cadeia LEFT_* precisa ficar em +X (média {})",
             left_sum / left_count as f32
         );
         assert!(
-            right_sum / right_count as f32 < 0.0,
+            right_sum / (right_count as f32) < 0.0,
             "cadeia RIGHT_* precisa ficar em −X (média {})",
             right_sum / right_count as f32
         );
@@ -901,7 +901,7 @@ mod tests {
         assert_eq!(summary.unmapped_vertices, mesh.vertices.len() - 10);
         assert!(!summary.is_complete());
         for vertex in mesh.vertices.iter().take(10) {
-            assert_ne!(vertex.joints[0] as usize, joints::ROOT);
+            assert_ne!(u32::from(vertex.joints[0]), joints::ROOT);
         }
         for vertex in mesh.vertices.iter().skip(10) {
             assert_eq!(vertex.joints[0], joints::HIPS as u16);
